requirements.txt
numpy
gradio
Pillow
rembg
onnx
-i https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/
onnxruntime-gpu
tensorrt==8.6.0
torch
safetensors
einops
torchvision
psutil
transformers
ollama
timm
flash_attn
xformers
pandas
huggingface-hub
onnxruntime
scikit-learn
SentencePiece
jsonformer